requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.0
pyparsing==2.4.7
PyPrind==2.11.2
python-dateutil==2.8.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
